% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBI_CE.R
\docType{data}
\name{DBI_CE}
\alias{DBI_CE}
\title{Central European checklist with DBI values}
\format{
Dataframe
}
\usage{
DBI_CE
}
\description{
Outdated checklist; see DBI_CEC for the updated checklist. Checklist of all
Central European odonate species with assigned DBI values according to Dolný et al. 2010;
contains 68 species (73 species in total, including DBI with NA values).
}
\references{
Harabiš, F., & Dolný, A. (2010). Využití vážek jako environmentálních indikátorů.
In A. Dolný & F. Harabiš (Eds.), Vážky 2010. Sborník referátů XIII. celostátního semináře
odonatologů v Podyjí (pp. 109–116). ZO ČSOP.
}
\author{
Harabiš et al., 2010
}
\keyword{datasets}
