% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBI_CEC.R
\docType{data}
\name{DBI_CEC}
\alias{DBI_CEC}
\title{Checklist of Odonate Species with DBI Values for Seven Central European Countries}
\format{
Dataframe
}
\usage{
DBI_CEC
}
\description{
A checklist of all Central European odonate species, including their assigned
DBI values, for seven Central European countries: Austria (AU), Czechia (CZ), Germany (DE),
Poland (PL), Slovakia (SK), Slovenia (SLO), and Switzerland (CHE), based on Bílková et al.
(submitted). The list comprises 88 species, some of which have DBI values listed as NA.
}
\references{
Bílková E., Šigutová H., Pyszko P., Prieložná V., & Dolný A. (submitted). Adopting
a country-specific Dragonfly Biotic Index: framework for seven Central European countries and
transboundary pattern analysis.
}
\author{
Bílková et al. (submitted)
}
\keyword{datasets}
