% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadData.R
\name{LoadData}
\alias{LoadData}
\title{Data uploading function}
\usage{
LoadData()
}
\value{
A data frame uploaded by the user containing a list of taxa in the first column, and abundance or presence/absence data in following columns, with sample names in the column header.
}
\description{
The function allows for simple loading of user data and converting NA values into null.
}
\examples{
Uploading data containing species of Central Europe. Saved as "Europe".

Europe<-LoadData()

Uploading data containing species of South Africa. Saved as "Africa".

Africa<-LoadData()
}
