% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadDBI.R
\name{LoadDBI}
\alias{LoadDBI}
\title{Uploading function for user-defined DBI checklist}
\usage{
LoadDBI(attrib = T)
}
\arguments{
\item{attrib}{}
}
\description{
Uploading function for user-defined DBI checklist
}
\examples{
# Uploading user-defind DBI checklist. Saved as DBI_MD.

DBI_MD<-LoadDBI()
}
