% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermDBI.R
\name{PermDBI}
\alias{PermDBI}
\title{Permutational DBI and potential for dragonfly communities}
\usage{
PermDBI(df, DBI_val, DBI_UD, NAval = T, sim = 10000, plot = F)
}
\arguments{
\item{plot}{}
}
\description{
Permutational DBI and potential for dragonfly communities
}
\examples{
# Calculate the permutational DBI potential for the Highway stormwater and control ponds dataset, comparing the given community to 10 000 randomly assembled communities.

Perm(DBI_Data, DBI_val = "CE", sim = 10000)
}
