% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PotDBI.R
\name{PotDBI}
\alias{PotDBI}
\title{DBI potential for dragonfly communities}
\usage{
PotDBI(df, DBI_val, DBI_UD, NAval = F, sim = 10000, plot = T)
}
\arguments{
\item{plot}{}
}
\description{
DBI potential for dragonfly communities
}
