% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PotDBI.R
\name{PotDBI_C}
\alias{PotDBI_C}
\title{DBI potential for dragonfly communities}
\usage{
PotDBI_C(df, DBI_val, DBI_UD, NAval = F, sim = 10000, plot = T)
}
\arguments{
\item{arrow}{}
}
\description{
DBI potential for dragonfly communities
}
