% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stormwaters.R
\docType{data}
\name{Stormwaters}
\alias{Stormwaters}
\title{Highway stormwater and control ponds dataset}
\format{
Dataframe
}
\usage{
Stormwaters
}
\description{
The dataset is derived from the research project of Aleš Dolný, which was used in
the publication: Highway stormwater ponds as islands of Odonata diversity in an agricultural
landscape (Šigutová et al., 2022).
}
\references{
Šigutová, H., Pyszko, P., Valušák, J., Dolný, A., 2022. Highway stormwater ponds as
islands of Odonata diversity in an agricultural landscape. Sci. Total Environ. 837, 155774.
https://doi.org/10.1016/j.scitotenv.2022.155774.
}
\author{
Šigutová et al., 2022
}
\keyword{datasets}
