% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateDBI.R
\name{CalculateDBI}
\alias{CalculateDBI}
\title{Main function calculating set of DBI values for dragonfly communities}
\usage{
CalculateDBI(df, DBI_val, DBI_UD, NAval = F, sim = 10000)
}
\arguments{
\item{plot}{}
}
\description{
Main function calculating set of DBI values for dragonfly communities
}
\examples{
# Calculate set of DBI values for the Highway stormwater and control ponds dataset

CalculateDBI(DBI_Data, DBI_val = "CE", plot = T)
}
