% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateDBI.R
\name{CalculateDBI}
\alias{CalculateDBI}
\title{Main function calculating set of DBI score types  for odonate communities}
\usage{
CalculateDBI(df, DBI_val, DBI_UD, NAval=F, sim=10000)
}
\arguments{
\item{df}{A data frame created by ‘UniteData’ function, containing a list of taxa in the first
column, followed by the columns with the values of distribution, threat and sensitivity
subindices of DBI (in case of using a pre-set checklist via ‘UniteData’ function), column of
DBI values, and columns of abundances with sample names in the rows.}

\item{DBI_val}{Indicates the checklist to be used for comparison. ‘SA’ refers to the South
African checklist with DBI values (Samways et al., 2016). ‘AU’ refers to the Austrian checklist,
‘CHE’ to the Swiss checklist, ‘CZ’ to the Czech checklist, ‘DE’ to the German checklist, ‘PL’ to
the Polish checklist, ‘SK’ to the Slovak checklist, and ‘SLO’ to the Slovenian checklist, all with
DBI values (Bílková et al., submitted). ‘UD’ refers to a user-defined or uploaded checklist.}

\item{DBI_UD}{In case that ‘UD’ is defined for the type, the name of user-loaded data frame
should be specified here.}

\item{NAval}{Logical, true in case of NA values in user-defined checklist.}

\item{sim}{The number of simulations identifies how many permutations should be made to
randomly assemble communities with the same species richness as the community in question.
The probability weight for each DBI is set as 2^-DBI, i.e., a species with a DBI higher by one
unit has half the probability of being selected into a random community than a species with a
lower DBI value.}
}
\value{
A data frame consisting of the columns of score types (sum of DBI, mean of DBI, DBI
potential, real DBI potential, and permutational DBI potential) with samples in rows.
}
\description{
#' @description Calculates sum of DBI, mean of DBI, DBI potential, real DBI potential, and
permutational DBI potential for odonate community samples.
}
\examples{
# For this function, you must have a data frame created by "UniteData".

# Unification of Highway stormwater and control ponds dataset and Czech checklist with DBI
# values. Saved as "StormwatersDBI".

\dontrun{
StormwatersDBI <- UniteData(Stormwaters, DBI_val = "CZ")

# "StormwatersDBI" is then the input of the “CalculateDBI” function.

# Calculate set of DBI score types for the Highway stormwater and control ponds dataset.
# Saved as "StormwatersCAL".

StormwatersCAL <- CalculateDBI(StormwatersDBI, DBI_val = "CZ")

# Then you can filter the specific calculation (column).

StormwatersCAL$PermDBI
}

\dontrun{
# Calculate set of DBI score types for species from South Africa. Saved as "AfricaCAL".

AfricaCAL <- CalculateDBI(AfricaDBI, DBI_val = "SA")

# Calculate set of DBI score types for species from Austria. Saved as "AustriaCAL".

AustriaCAL <- CalculateDBI(Austria, DBI_val = "AU")

# Calculate set of DBI score types for species from user data. In this case the input of
# CalculateDBI" function is also users DBI checklist ("DBI_UD"). Saved as "UserDataCAL".

UserDataCAL <- CalculateDBI(UserDataDBI, DBI_val = "UD", DBI_UD)
}
}
