% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBI_CE.R
\docType{data}
\name{DBI_CE}
\alias{DBI_CE}
\title{Central European checklist with DBI values}
\format{
A data frame
}
\usage{
DBI_CE
}
\description{
Checklist of all Central European dragonfly species with assigned DBI values according to Dolný et al. 2012. Contains 68 species.
}
\references{
Dolný, A., Harabiš, F., Bárta, D., Lhota, S., Drozd, P., 2012. Aquatic insects indicate terrestrial habitat degradation: Changes in taxonomical structure and functional diversity of dragonflies in tropical rainforest of East Kalimantan. Trop. Zool. 25, 141–157.
}
\author{
Dolný et al., 2012
}
\keyword{datasets}
