% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBI_SA.R
\docType{data}
\name{DBI_SA}
\alias{DBI_SA}
\title{South African checklist with DBI values}
\format{
Dataframe
}
\usage{
DBI_SA
}
\description{
Checklist of all South African odonate  species with assigned DBI values according to Samways et al., 2016; contains 167 species.
}
\references{
Samways, M.J., Simaika, J.P., 2016. Manual of freshwater assessment for South Africa: Dragonfly Biotic Index. Suricata 2. South African National Biodiversity Institute, Pretoria.
}
\author{
Samways et al., 2016
}
\keyword{datasets}
