% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadData.R
\name{LoadData}
\alias{LoadData}
\title{Data uploading function}
\usage{
LoadData(attrib = T, na2null = T, abunance = T)
}
\arguments{
\item{na2null}{}
}
\description{
Data uploading function
}
\examples{
# Uploading data containing species of Central Europe. Saved as x_CE.

x_CE<-LoadData()


# Uploading data containing species of South Africa. Saved as x_SA.

x_SA<-LoadData()
}
