% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadDBI.R
\name{LoadDBI}
\alias{LoadDBI}
\title{Uploading function for user-defined DBI checklist}
\usage{
LoadDBI()
}
\description{
The function allows for uploading a user-defined checklist of odonates with a user-defined set of DBI values.
}
\examples{
# Uploading user-defined DBI checklist. Saved as "DBI_UD".

DBI_UD<-LoadDBI()
}
