% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermDBI.R
\name{PermDBI}
\alias{PermDBI}
\title{Permutational DBI potential for odonate communities}
\usage{
PermDBI(df, DBI_val, DBI_UD, NAval = T, sim = 10000, plot = F)
}
\arguments{
\item{df}{A data frame created by ‘UniteData’ function, containing a list of taxa in the first
column, followed by the columns with the values of distribution, threat and sensitivity subindices
(in case of using a pre-set checklist via ‘UniteData’ function), column of DBI values, and columns
of abundances with sample names in the rows.}

\item{DBI_val}{Indicates the checklist to be used for comparison. ‘SA’ refers to the South African
checklist with DBI values (Samways et al., 2016). ‘AU’ refers to the Austrian checklist, ‘CHE’ to
the Swiss checklist, ‘CZ’ to the Czech checklist, ‘DE’ to the German checklist, ‘PL’ to the Polish
checklist, ‘SK’ to the Slovak checklist, and ‘SLO’ to the Slovenian checklist, all with DBI values
(Bílková et al., submitted). ‘UD’ refers to a user-defined or uploaded checklist.}

\item{DBI_UD}{In case that ‘UD’ is defined for the type, the name of user-loaded data frame should
be specified here.}

\item{NAval}{Logical, true in case of NA values in user-defined checklist.}

\item{sim}{The number of simulations identifies how many permutations should be made to randomly
assemble communities with the same species richness as the community in question.}

\item{plot}{Should a histogram for randomly assembled communities be plotted for each compared
community, completed with a vertical abline representing the given community? By default, the plot
is not rendered.}
}
\value{
A data frame consisting of a column of index values with samples in rows.
}
\description{
Calculates permutational DBI potential for odonate community samples.
}
\examples{
# For this function, you must have a data frame created by "UniteData".

# Unification of Highway stormwater and control ponds dataset and Czech checklist with DBI values.
# Saved as "StormwatersDBI".

\dontrun{
StormwatersDBI <- UniteData(Stormwaters, DBI_val = "CZ")

# "StormwatersDBI" is then the input of the “CalculateDBI” function.

# Calculate the permutational DBI potential for the Highway stormwater and control ponds dataset,
# comparing the given community to 10,000 randomly assembled communities.

PermDBI(StormwaterDBI, DBI_val = "CZ")

# Calculate the permutational DBI potential for the user data with own DBI checklist and plotted
# results.

PermDBI(UserDataDBI, DBI_val = "UD", DBI_UD, plot = T)
}

}
