% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PotDBI.R
\name{PotDBI}
\alias{PotDBI}
\title{DBI potential for dragonfly communities}
\usage{
PotDBI(df, DBI_val, DBI_UD, type = "def", NAval = F, plot = F)
}
\arguments{
\item{plot}{}
}
\description{
DBI potential for dragonfly communities
}
