% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StandardDBI.R
\name{StandardDBI}
\alias{StandardDBI}
\title{Standard DBI values for dragonfly communities}
\usage{
StandardDBI(df, UD = F, type, data = "DBI")
}
\arguments{
\item{data}{}
}
\description{
Standard DBI values for dragonfly communities
}
\examples{
# Calculate the sum DBI index for the Highway stormwater and control ponds dataset.

StandardDBI(DBI_Data, type = "sum")

# Calculate the mean DBI index for the Highway stormwater and control ponds dataset.

StandardDBI(DBI_Data, type = "mean")
}
