% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniteData.R
\name{UniteData}
\alias{UniteData}
\title{Unification of users’ dataset and DBI data}
\usage{
UniteData(df, DBI_val, DBI_UD)
}
\arguments{
\item{DBI_UD}{}
}
\description{
Unification of users’ dataset and DBI data
}
\examples{
# Unification of Highway stormwater and control ponds dataset and Central European checklist with DBI values. Saved as DBI_Data.

DBI_Data<-UniteData(Data, DBI_val = "CE")
}
